+++ apps/etterlatte-behandling/src/main/kotlin/Contekst.kt
@@ -4,13 +4,13 @@ import kotlinx.coroutines.runBlocking
 import no.nav.etterlatte.behandling.EnhetService
 import no.nav.etterlatte.common.Enheter
 import no.nav.etterlatte.libs.ktor.AZURE_ISSUER
+import no.nav.etterlatte.tilgangsstyring.SaksbehandlerMedRoller
 import no.nav.etterlatte.token.Bruker
 import no.nav.etterlatte.token.Saksbehandler
 import no.nav.etterlatte.token.SystemBruker
 import no.nav.security.token.support.core.context.TokenValidationContext
 import no.nav.security.token.support.v2.TokenValidationContextPrincipal
 import tilgangsstyring.AzureGroup
-import tilgangsstyring.SaksbehandlerMedRoller
 import java.sql.Connection
 
 object Kontekst : ThreadLocal<Context>()
